ars-update: another try at the 'no arch-specific dirs' problem
As APT's lists are all keyed by binary arch, we need to know for each
distribution which architectures are used.  However, a minimal dist
might only have arch-independent packages.  In that case, build lists
as if every arch seen for any dist was used.  Otherwise, if even a
single one is seen, then only lists for those are written.

We might still end up leaving stale lists behind, so there is still
something of a FIXME implicit in this change.